componentizacion y refactorizacion de reviews
